requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==11.1.0
lxml==5.3.0
playwright==1.51.0
python-dotenv==1.0.1
